feat: support 128-bit integers in inequalities (PROOF-548) (#357)
* fill in binary proof

* fill in sign expr

* fill in sign expr

* rework verification builder

* rework const verification

* rework

* refactor

* add tests

* add tests

* fill in tests

* add tests

* doc

* work on const absolute verification

* constant abs decomposition

* support constant absolute

* add test

* fix tests

* add tests

* fill in general sign expr support

* fill in sign expr

* fill in sign expr

* fix clippy

* fix verification

* add tests

* work on lte expr testing

* fill in random test

* rework inequality

* rework

* drop old version

* document

* support 128 bit

+++ crates/proofs/src/sql/ast/sign_expr_test.rs
@@ -47,7 +47,7 @@ fn prover_evaluation_generates_the_bit_distribution_of_a_negative_constant_colum
 #[test]
 fn count_fails_if_a_bit_distribution_is_out_of_range() {
     let dists = [BitDistribution::new(&[
-        ArkScalar::from(3) * ArkScalar::from(u64::MAX)
+        ArkScalar::from(3) * ArkScalar::from(u128::MAX)
     ])];
     let mut builder = CountBuilder::new(&dists);
     assert!(count_sign(&mut builder).is_err());
